test(esplora): move esplora tests into src files
Since we want to keep these methods private.
